Break PDB loader into modules
